requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.33.2
torch==2.0.1
beautifulsoup4==4.12.2
requests==2.31.0
yfinance==0.2.28
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sentence-transformers==2.2.2 